Begin rearranging the graph
