+++ oc/compiler/graph_analyzer/graph_analyzer.h
@@ -12,5 +12,24 @@
 #include "../cfg/cfg.h"
 
 
+/**
+ * We will calculate:
+ *  1.) Dominator Sets
+ *  2.) Dominator Trees
+ *  3.) Dominance Frontiers
+ *  4.) Postdominator sets
+ *  5.) Reverse Dominance frontiers
+ *  6.) Reverse post order traversals
+ *
+ * For every block in the given function
+ */
+void calculate_all_control_flow_relations_for_function(basic_block_t* function_entry, dynamic_array_t* function_blocks);
+
+/**
+ * Destroy all old control relations in anticipation of new ones coming in. This
+ * operates on a per-function level
+ */
+void cleanup_all_control_relations(dynamic_array_t* function_blocks);
+
 
 #endif /* GRAPH_ANALYZER_H */
